# Fix MemoryCache Compact performance.