- Added a benchmark to validate the performance gains
- As the amount of elements gets larger in the cache the gains get significantly higher too because of how the old "OrderBy" & "ElementAt" calls work. Aka, every ElementAt call would re-calculate the entirety of the OrderBy.

Before:
|                                          Method |     Mean |    Error |   StdDev |  Op/s |     Gen 0 |    Gen 1 | Gen 2 | Allocated |
|------------------------------------------------ |---------:|---------:|---------:|------:|----------:|---------:|------:|----------:|
| 'MemoryCache Set performance with limited size' | 17.47 ms | 0.137 ms | 0.114 ms | 57.23 | 2281.2500 | 156.2500 |     - |  13.78 MB |

After:
|                                          Method |     Mean |    Error |   StdDev |    Op/s |   Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------------------------------------ |---------:|---------:|---------:|--------:|--------:|-------:|------:|----------:|
| 'MemoryCache Set performance with limited size' | 695.9 us | 13.34 us | 14.27 us | 1,437.0 | 42.9688 | 3.9063 |     - |    266 KB |

**Gains:** 25x speed improvement

/cc @ToddGrun I forget, did you all utilize the MemoryCache in places?

Fixes dotnet/aspnetcore#29500