@NTaylorMullen 2021-01-22
> Good eye! How did you end up finding this to be a bottleneck? In that investigation did you notice any other candidates for optimization (and do we have bugs for them).

Just lots of experience with IEnumerable's being dangerous is all 😄. Because IEnumerable's are lazy and uncached you can get into some (O)n^2 scenarios super easily unless you force evaluation (ToArray).